timer aka score
for rotating pong i added a timer. to know when the player is not playing anymore. i want it to be something like "you never know when game over is going to happn".

Game over happen when the playerdoes not touch the ball with the paddlefor 60 seconds

all made a function for teh game over screen